GenerarBack/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
requests==2.28.1
python-dotenv==0.21.0
Flask-Cors==3.0.10